Move "src/components/smart-grid/internal/smart-grid-virtual-scroller/" folder to "src/components/"
