+++ src/smart-grid-virtual-scroller/cells-in-viewport-are-rendered-and-visible.ts
@@ -1,4 +1,4 @@
-import { inBetween } from "../../../../common/utils";
+import { inBetween } from "../common/utils";
 import {
   cellIsRendered,
   getSmartCells,
