+++ src/smart-grid-virtual-scroller/types.ts
@@ -1,4 +1,4 @@
-import { SmartGridModel } from "../../types";
+import { SmartGridModel } from "../components/smart-grid/types";
 
 export type VirtualScrollVirtualItems = {
   virtualItems: SmartGridModel;
