+++ src/smart-grid-virtual-scroller/virtual-scroller.scss
